- project settings update - desktop compiler float bug fix

+++ openGL empty template/assets/fragmentshader.fs
@@ -6,15 +6,15 @@ uniform float checkerCount;
 uniform vec2 mousepos;
 
 void main (void) {
-    float checkerSize = 1f/checkerCount;
+    float checkerSize = (float)1/checkerCount;
     float brightness = 0.0f;
     float spotrange = 150.0f;
-    float spotbrightness = 1f;
+    float spotbrightness = 1;
     float ambientIntensity = 0.1f;
     vec4 ambientColor = vec4(1,0,0,1);
     vec4 spotcolor = vec4(1,1,1,1);
 
-    float dist = distance (vec3(mousepos.x, mousepos.y, 0f), vec3(gl_FragCoord));
+    float dist = distance (vec3(mousepos.x, mousepos.y, 0), vec3(gl_FragCoord));
 
     brightness = 1- dist/spotrange;
     if(brightness <0){
